requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

# WSGI Handler
gunicorn==20.1.0

# Static and Media Storage
django-storages[s3]==1.14.2

# Email backends
django-anymail[mailgun]==10.2

# Monitoring
sentry-sdk==1.17.0

# Cache and Job Queue
redis==4.5.5
celery==5.3.6
flower==2.0.1
